requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
opencv-python==4.8.1.78
numpy==1.26.2
torch==2.1.1
transformers==4.35.2
librosa==0.10.1
soundfile==0.12.1
mediapipe==0.10.8
streamlit==1.28.2
pandas==2.1.3
altair==5.1.2
python-dotenv==1.0.0